requirements.txt
numpy
itk-elastix
itkwidgets
zarr
dask
dask-image
cupy-cuda12x
structure-tensor
monai
matplotlib==3.9.0
torchio
opencv-python
qim3d
hydra-core
invoke
scikit-learn
scikit-image
matplotlib
napari
pipreqs
typer
fastapi
uvicorn
coverage
pytest
ruff
python-multipart
httpx
locust
netron
bentoml
evidently
pandas
h5py